requirements.txt
-i https://pypi.org/simple
requests==2.27.1
pandas==1.4.1
numpy==1.22.3
python-dotenv==0.19.2
APScheduler==3.7.0